Refactored types declaration
To have the type declarations and interfaces available in all the environment-specific modules, they were moved from oidc-client-dpop-browser back to the core module. That means that the oidc-client-dpop-browser depends on a packages that was initially meant to be solid-specific, making it less of a "pure OIDC" library, but the cost is smaller than having an independant package exporting pure OIDC types at this stage.
