+++ packages/core/src/index.ts
@@ -45,8 +45,14 @@ export {
   ISessionInfoManagerOptions,
 } from "./sessionInfo/ISessionInfoManager";
 
-export { default as IIssuerConfigFetcher } from "./login/oidc/IIssuerConfigFetcher";
-export { default as issuerConfigSchema } from "./login/oidc/issuerConfigSchema";
+export { IIssuerConfigFetcher } from "./login/oidc/IIssuerConfigFetcher";
+export { IIssuerConfig } from "./login/oidc/IIssuerConfig";
+export {
+  IClientRegistrar,
+  IClientRegistrarOptions,
+} from "./login/oidc/IClientRegistrar";
+export { IClient } from "./login/oidc/IClient";
+export { issuerConfigSchema } from "./login/oidc/issuerConfigSchema";
 
 // Storage.
 export { default as IStorage } from "./storage/IStorage";
